Move UnionDataset to dataset_utils.py
